# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Patna Realty Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /inquiries:
    get:
      operationId: listInquiries
      tags: [inquiries]
      summary: List buyer inquiries
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, closed]
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Submit a buyer inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryBody"
      responses:
        "201":
          description: Inquiry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /inquiries/{inquiryId}:
    put:
      operationId: updateInquiry
      tags: [inquiries]
      summary: Update inquiry status
      parameters:
        - name: inquiryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInquiryBody"
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
...
  /dashboard/rent-roll:
    get:
      operationId: getRentRoll
      tags: [dashboard]
      summary: Get rent roll report
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: YYYY-MM format
      responses:
        "200":
          description: Rent roll
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentRollEntry"
  /dashboard/collection-report:
    get:
      operationId: getCollectionReport
      tags: [dashboard]
      summary: Payment collection report by period
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Collection report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionReport"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, tenant, shopkeeper, buyer]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [tenant, shopkeeper, buyer]
      required: [name, email, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    Property:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [flat, shop, grocery_store, market]
        bhk:
          type: ["integer", "null"]
          description: BHK count, applicable only for flats
        sizeSqft:
          type: number
        floor:
          type: ["integer", "null"]
        blockOrSector:
          type: ["string", "null"]
        address:
          type: string
        locality:
          type: string
        landmark:
          type: ["string", "null"]
        rentPrice:
          type: ["number", "null"]
        salePriceINR:
          type: ["number", "null"]
        rentPeriod:
          type: ["string", "null"]
          enum: [daily, weekly, monthly, yearly, null]
        purpose:
          type: string
          enum: [rent, sale, both]
        status:
          type: string
          enum: [available, occupied, for_sale]
        description:
          type: ["string", "null"]
        photos:
          type: array
          items:
            type: string
        businessType:
          type: ["string", "null"]
          description: For shops/grocery stores
        amenities:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, type, sizeSqft, address, locality, purpose, status, photos, amenities, createdAt, updatedAt]
...
    Inquiry:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        property:
          oneOf:
            - $ref: "#/components/schemas/Property"
            - type: "null"
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [new, contacted, closed]
        createdAt:
          type: string
          format: date-time
      required: [id, propertyId, name, email, phone, message, status, createdAt]
    CreateInquiryBody:
      type: object
      properties:
        propertyId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
      required: [propertyId, name, email, phone, message]
    UpdateInquiryBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, closed]
      required: [status]
...
    RentRollEntry:
      type: object
      properties:
        tenantName:
          type: string
        propertyName:
          type: string
        propertyType:
          type: string
        unitNumber:
          type: ["string", "null"]
        rentAmount:
          type: number
        rentPeriod:
          type: string
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, partial, overdue]
        amountPaid:
          type: ["number", "null"]
        balance:
          type: ["number", "null"]
      required: [tenantName, propertyName, propertyType, rentAmount, rentPeriod, dueDate, status]
    CollectionReport:
      type: object
      properties:
        period:
          type: string
        totalExpected:
          type: number
        totalCollected:
          type: number
        totalPending:
          type: number
        collectionRate:
          type: number
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/CollectionBreakdownItem"
      required: [period, totalExpected, totalCollected, totalPending, collectionRate, breakdown]
    CollectionBreakdownItem:
      type: object
      properties:
        label:
          type: string
        expected:
          type: number
        collected:
          type: number
      required: [label, expected, collected]
...